+++ src/index.ts
@@ -1,2 +1,1 @@
-export { RapidServer } from "./server/RapidServer"
-export { getRapidObjectClass } from "./builder/rapid-class.builder"
+export { RapidServer } from "./RapidServer"