+++ src/types/allowed-http-methods.ts
@@ -1,1 +1,0 @@
-export type AllowedHttpMethods = "GET" | "POST" | "PUT" | "DELETE" | "PATCH";