+++ src/types/openapi-spec.d.ts
@@ -1,0 +1,19 @@
+export type OpenApiSpec = {
+	openapi: string,
+	servers: {
+		url: string,
+		description: string,
+	}[],
+	info: {
+		title: string,
+		version: string,
+	},
+	tags: {
+		name: string,
+		description: string,
+	}[],
+	paths: {},
+	components: {
+		schemas: {},
+	}
+}