+++ src/types/rapid-endpoints.ts
@@ -1,8 +1,0 @@
-import {RapidHttpMethods} from "./rapid-http-methods";
-
-export type RapidEndpoint = {
-    name: string,
-    methods: RapidHttpMethods[],
-    object: string,
-    hasId: boolean
-}